split terminal into another project and add Avalonia UI project

+++ iCUE-ReverseEngineer/Icue/Data/IcueLed.cs
@@ -2,7 +2,7 @@
 
 public class IcueLed
 {
-    public int Width { get; set; }
+    public double Width { get; set; }
     public int LedId { get; set; }
     public double Top { get; set; }
     public double Left { get; set; }
